+++ src/literalizer/languages/cpp.py
@@ -23,6 +23,7 @@ from literalizer._formatters.format_entries import (
     braced_dict_entry,
     format_bytes_base64,
     format_bytes_hex,
+    passthrough_call_arg,
     passthrough_sequence_entry,
     passthrough_set_entry,
     variable_formatter,
@@ -1363,6 +1364,11 @@ class Cpp(metaclass=LanguageCls):
             date_type=self._cpp_date_type,
             datetime_type=self._cpp_datetime_type,
         )
+
+    @cached_property
+    def format_call_arg(self) -> Callable[[Value, str], str]:
+        """Wrap a formatted call argument."""
+        return passthrough_call_arg
 
     @cached_property
     def heterogeneous_behavior(self) -> HeterogeneousBehavior: