+++ tests/integration/cases/call_per_element_false/ObjectiveC_call.m
@@ -1,7 +1,5 @@
 #import <Foundation/Foundation.h>
-#pragma clang diagnostic ignored "-Wstrict-prototypes"
-#pragma clang diagnostic ignored "-Wdeprecated-non-prototype"
-void process();
+void process(id);
 void check_(void) {
-process(1);
+process(@(1));
 }