+++ tests/integration/cases/call_scalar_args/ObjectiveC_call.m
@@ -1,9 +1,7 @@
 #import <Foundation/Foundation.h>
-#pragma clang diagnostic ignored "-Wstrict-prototypes"
-#pragma clang diagnostic ignored "-Wdeprecated-non-prototype"
-void process();
+void process(id);
 void check_(void) {
 process(@"hello");
-process(42);
+process(@(42));
 process(@YES);
 }